+++ src/main.rs
@@ -1,6 +1,0 @@
-// This is free and unencumbered software released into the public domain.
-
-fn main() -> Result<(), ()> {
-    println!("asimov-huggingface-cataloger"); // TODO
-    Ok(())
-}